requirements.txt
bidict==0.21.4
click==8.0.3
Flask==2.0.2
Flask-SocketIO==5.1.1
h11==0.12.0
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
python-decouple==3.5
python-engineio==4.3.0
python-socketio==5.4.1
simple-websocket==0.4.0
Werkzeug==2.0.2
wsproto==1.0.0
